backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-multipart==0.0.6
httpx==0.24.1
numpy==1.26.3
opencv-python-headless==4.8.1.78
Pillow==10.0.1
tqdm==4.66.1
pydantic==1.10.12
requests==2.32.1
# optional (enable when using torch/lightweight models)
torch==2.2.0  # or install appropriate for your CUDA
torchvision==0.17.0
timm==0.9.2
ultralytics==8.0.131  # YOLOv8 if you want detector
